*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3783 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/vars_util.c
@@ -869,7 +869,7 @@ Var *vars_div_of_int_var( Var *v1, Var *v2 )
 	size_t elems;
 	long *lp;
 	long *v2_lpnt;
-	long *dp;
+	double *dp;
 	double *v2_dpnt;
 
 
@@ -1175,7 +1175,7 @@ Var *vars_mod_of_int_var( Var *v1, Var *v2 )
 	size_t elems;
 	long *lp;
 	long *v2_lpnt;
-	long *dp;
+	double *dp;
 	double *v2_dpnt;
